リファクタリング

+++ StarLight_Line/GameLib/JoyconManager/Joycon/hid/hid.cpp
@@ -636,7 +636,7 @@ extern "C" {
 			length = dev->output_report_length;
 		}
 
-		res = WriteFile(dev->device_handle, buf, length, NULL, &ol);
+		res = WriteFile(dev->device_handle, buf, static_cast<DWORD>(length), NULL, &ol);
 
 		if (!res) {
 			if (GetLastError() != ERROR_IO_PENDING) {
@@ -679,7 +679,7 @@ extern "C" {
 			dev->read_pending = TRUE;
 			memset(dev->read_buf, 0, dev->input_report_length);
 			ResetEvent(ev);
-			res = ReadFile(dev->device_handle, dev->read_buf, dev->input_report_length, &bytes_read, &dev->ol);
+			res = ReadFile(dev->device_handle, dev->read_buf, static_cast<DWORD>(dev->input_report_length), &bytes_read, &dev->ol);
 
 			if (!res) {
 				if (GetLastError() != ERROR_IO_PENDING) {
@@ -733,7 +733,7 @@ extern "C" {
 			return -1;
 		}
 
-		return copy_len;
+		return static_cast<int>(copy_len);
 	}
 
 	int HID_API_EXPORT HID_API_CALL hid_read(hid_device *dev, unsigned char *data, size_t length)
@@ -749,13 +749,13 @@ extern "C" {
 
 	int HID_API_EXPORT HID_API_CALL hid_send_feature_report(hid_device *dev, const unsigned char *data, size_t length)
 	{
-		BOOL res = HidD_SetFeature(dev->device_handle, (PVOID)data, length);
+		BOOL res = HidD_SetFeature(dev->device_handle, (PVOID)data, static_cast<ULONG>(length));
 		if (!res) {
 			register_error(dev, "HidD_SetFeature");
 			return -1;
 		}
 
-		return length;
+		return static_cast<int>(length);
 	}
 
 
@@ -777,8 +777,8 @@ extern "C" {
 
 		res = DeviceIoControl(dev->device_handle,
 			IOCTL_HID_GET_FEATURE,
-			data, length,
-			data, length,
+			data, static_cast<DWORD>(length),
+			data, static_cast<DWORD>(length),
 			&bytes_returned, &ol);
 
 		if (!res) {
@@ -819,7 +819,7 @@ extern "C" {
 	{
 		BOOL res;
 
-		res = HidD_GetManufacturerString(dev->device_handle, string, sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS));
+		res = HidD_GetManufacturerString(dev->device_handle, string, static_cast<ULONG>(sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS)));
 		if (!res) {
 			register_error(dev, "HidD_GetManufacturerString");
 			return -1;
@@ -832,7 +832,7 @@ extern "C" {
 	{
 		BOOL res;
 
-		res = HidD_GetProductString(dev->device_handle, string, sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS));
+		res = HidD_GetProductString(dev->device_handle, string, static_cast<ULONG>(sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS)));
 		if (!res) {
 			register_error(dev, "HidD_GetProductString");
 			return -1;
@@ -845,7 +845,7 @@ extern "C" {
 	{
 		BOOL res;
 
-		res = HidD_GetSerialNumberString(dev->device_handle, string, sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS));
+		res = HidD_GetSerialNumberString(dev->device_handle, string, static_cast<ULONG>(sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS)));
 		if (!res) {
 			register_error(dev, "HidD_GetSerialNumberString");
 			return -1;
@@ -858,7 +858,7 @@ extern "C" {
 	{
 		BOOL res;
 
-		res = HidD_GetIndexedString(dev->device_handle, string_index, string, sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS));
+		res = HidD_GetIndexedString(dev->device_handle, string_index, string, static_cast<ULONG>(sizeof(wchar_t) * MIN(maxlen, MAX_STRING_WCHARS)));
 		if (!res) {
 			register_error(dev, "HidD_GetIndexedString");
 			return -1;
